+++ src/nqdc/_links.py
@@ -1,4 +1,6 @@
 """Extracting all URLs from an article XML."""
+import pathlib
+
 import pandas as pd
 from lxml import etree
 
@@ -16,7 +18,9 @@ class LinkExtractor(Extractor):
     fields = ("pmcid", "ext-link-type", "href")
     name = "links"
 
-    def extract(self, article: etree.ElementTree) -> pd.DataFrame:
+    def extract(
+        self, article: etree.ElementTree, article_dir: pathlib.Path
+    ) -> pd.DataFrame:
         pmcid = _utils.get_pmcid(article)
         all_links = []
         xlink = "http://www.w3.org/1999/xlink"